LF-22: db and services refactoring
