+++ app/handler/activity/delete_activity.py
@@ -2,9 +2,9 @@ from telebot import TeleBot
 
 from app.config import msg
 from app.handler.general import TelegramMessageHandler, TelegramCallbackHandler, MessageMeta, CallbackMeta
-from app.service import ActivityService
-from app.util import markup
-from app.util.markup import EMPTY_VOTE_RESULT
+from app.service.activity import ActivityService
+from app.service import markup
+from app.service.markup import EMPTY_VOTE_RESULT
 
 
 class DeleteActivityHandler(TelegramMessageHandler):
