+++ app/handler/statistics/statistics.py
@@ -2,7 +2,7 @@ from telebot import TeleBot
 
 from app.config import msg
 from app.handler.general import TelegramMessageHandler, MessageMeta
-from app.service import StatisticsService
+from app.service.statistics import StatisticsService
 
 
 class StatisticsPostAnswerHandler(TelegramMessageHandler):
