--- app/validator.py
@@ -1,3 +1,0 @@
-def validate_username(username: str) -> tuple:
-    if username is None:
-        return "False"
